Refactor CEL code to the operator.
