+++ cmd/sink/filters/namespace_gvk.go
@@ -1,7 +1,7 @@
 // Copyright KubeArchive Authors
 // SPDX-License-Identifier: Apache-2.0
 
-package expr
+package filters
 
 import (
 	"fmt"
